backend/pyproject.toml
[project]
name = "beadsprite-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pillow>=10.1.0",
    "numpy>=1.26.0",
    "scikit-image>=0.22.0",
    "python-multipart>=0.0.6",
    "opencv-python>=4.11.0.86",
    "scipy>=1.16.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
